pyproject.toml
[project]
name = "cs423-color-segmentation"
version = "0.1.0"
description = "CS423/523 color-based segmentation project scaffold"
requires-python = ">=3.9"
dependencies = ["numpy>=1.26,<3.0", "pillow>=10.4,<12.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = ["node_modules"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
